requirements.txt
PyQt5>=5.15.6
websocket-client>=1.3.1
setuptools>=58.1.0
